# openapi.yml
openapi: 3.0.0
info:
  title: Trivia API
  description: API for a trivia game
  version: 1.0.0
servers:
  - url: http://localhost:8083
    description: Development server
paths:
  /rooms:
    get:
      summary: Get rooms (only public ones without a code)
      operationId: getRooms
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Rooms retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"

    post:
      summary: Create a room
      operationId: createRoom
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /rooms/{id}:
    get:
      summary: Get a room
      operationId: getRoom
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Room retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "401":
          description: Player not authenticated
        "403":
          description: Only players inside a private room can view it
        "404":
          description: Room not found

    delete:
      summary: Delete a room
      operationId: deleteRoom
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Room deleted
        "401":
          description: Player not authenticated
        "403":
          description: Only the host can delete a room
        "404":
          description: Room not found
        "409":
          description: Cannot delete a room which has a game
  /rooms/{id}/events:
    get:
      summary: Subscribe to server-sent events (SSEs) about the room
      operationId: subscribeToRoomEvents
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: SSE stream started
          content:
            text/event-stream:
              schema:
                type: string
        "404":
          description: Room not found
  /players:
    post:
      summary: Join a room as a player
      operationId: joinRoom
      parameters:
        - name: roomId
          in: query
          required: true
          schema:
            type: integer
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: username
          in: query
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Joined the room
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        "400":
          description: Invalid room code
        "404":
          description: Room not found

    get:
      summary: Get the players of a room
      operationId: getPlayers
      parameters:
        - name: roomId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Players retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"
        "404":
          description: Room not found
  /players/{id}:
    get:
      summary: Get a player
      operationId: getPlayer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Player retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "404":
          description: Player not found

    delete:
      summary: Remove a player from a room
      operationId: deletePlayer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Player deleted
        "401":
          description: Player not authenticated
        "403":
          description: Only the host can delete another player
        "404":
          description: Player not found
...
  /rounds:
    get:
      summary: Get the rounds of a game
      operationId: getRounds
      parameters:
        - name: gameId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rounds retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Round"
        "404":
          description: Game not found
  /rounds/{id}:
    get:
      summary: Get a round
      operationId: getRound
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Round retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Round"
        "404":
          description: Round not found
  /questions:
    get:
      summary: Get the questions of a round (correct answers hidden until the round ends)
      operationId: getQuestions
      parameters:
        - name: roundId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Questions retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
        "403":
          description: Round has not started yet
        "404":
          description: Round not found
...
components:
  schemas:
    Room:
      type: object
      properties:
        id:
          type: integer
        hostId:
          type: integer
          nullable: true # null if no host in the room
        gameId:
          type: integer
          nullable: true # null if no game created yet in the room
        createdAt:
          type: string
          format: date-time
        code:
          type: string
          nullable: true # null if room is public
...
    Player:
      type: object
      properties:
        id:
          type: integer
        roomId:
          type: integer
        teamId:
          type: integer
          nullable: true # null if player is not in a team
        username:
          type: string
...
    Round:
      type: object
      properties:
        id:
          type: integer
        gameId:
          type: integer
        createdAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
    Question:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        question:
          type: string
        mediaUrl:
          type: string
          nullable: true
        options:
          type: array
          items:
            type: string
        correctAnswers:
          type: array
          items:
            type: string
...